src/job_finder/requirements.txt
boto3
google-genai
google-search-results